refactor: package ui의 component 폴더 구조 변경
