+++ packages/ui/src/components/atoms/AccordionItem/index.module.scss
