+++ packages/ui/src/components/atoms/Badge/Badge.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import Badge from '.';
 
 const meta = {
-  title: 'UI/Badge',
+  title: 'atoms/Badge',
   component: Badge,
   parameters: {
     layout: 'centered',
