+++ packages/ui/src/components/index.ts
@@ -1,6 +1,6 @@
-export { default as AccordionItem } from './AccordionItem';
-export { default as Badge } from './Badge';
-export type { ButtonProps, ButtonSize } from './Button';
-export { default as Button } from './Button';
-export { Counter, default as CounterCard } from './CounterCard';
-export { default as SkillTag } from './SkillTag';
+export { default as AccordionItem } from './atoms/AccordionItem';
+export { default as Badge } from './atoms/Badge';
+export { default as SkillTag } from './atoms/SkillTag';
+export type { ButtonProps, ButtonSize } from './molecules/Button';
+export { default as Button } from './molecules/Button';
+export { Counter, default as CounterCard } from './molecules/CounterCard';
