+++ packages/ui/src/components/molecules/CounterCard/index.module.scss
